+++ Source/wsadphysics/RocketManPlayerController.h
@@ -5,6 +5,7 @@
 #include "CoreMinimal.h"
 #include "GameFramework/PlayerController.h"
 #include "PauseMenuWidget.h"
+#include "InputActionValue.h"
 #include "wsadphysicsCharacter.h"
 #include "RocketManPlayerController.generated.h"
 
@@ -27,13 +28,37 @@ protected:
 	UPROPERTY()
 	UPauseMenuWidget* PauseMenuWidget;
 
+	/** MappingContext */
+	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = Input, meta = (AllowPrivateAccess = "true"))
+	class UInputMappingContext* DefaultMappingContext;
+
+	/** Jump Input Action */
+	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = Input, meta = (AllowPrivateAccess = "true"))
+	class UInputAction* JumpAction;
+
+	/** Move Input Action */
+	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = Input, meta = (AllowPrivateAccess = "true"))
+	class UInputAction* MoveAction;
+
 
 	UFUNCTION()
 	void OpenPauseMenu();
+
+	UFUNCTION()
+	void StartThruster();
+	
+	UFUNCTION()
+	void StopThruster();
+	
+	UFUNCTION()
+	void Thrust(const float ThrustValue);
+	
+	UFUNCTION()
+	void Rotate(const FInputActionValue& Value);
 	
 public:
 	virtual void SetupInputComponent() override;
-	
+
 private:
 	UPROPERTY() FString DefaultMainMenuLevelName = "MainMenuMap";
